Rag_Module/requirements.txt
# Core LangChain ecosystem
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-classic>=0.1.0
langchain-text-splitters>=0.2.0

# Vector Store
langchain-chroma>=0.1.0
chromadb>=0.5.0

# Embeddings
langchain-huggingface>=0.0.3
sentence-transformers>=2.6.0
transformers>=4.40.0
torch>=2.2.0

# NVIDIA AI Endpoints
langchain-nvidia-ai-endpoints>=0.1.0

# Ranking / Reranking
scikit-learn>=1.3.0
rank-bm25>=0.2.2

# Utilities
pydantic>=2.0.0
numpy>=1.26.0
tqdm>=4.66.0

# Environment
python-dotenv>=1.0.0

# Job Queue & Caching
redis>=5.0.0
